# im-backoffice/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  errors.Response:
    properties:
      key:
        type: string
      message:
        type: string
    type: object
...
  inventoryCategory.InventoryCategoryDTO:
    properties:
      id:
        type: string
      inventory_category_name:
        type: string
    type: object
  inventoryCategory.InventoryCategoryReq:
    properties:
      inventory_category_name:
        type: string
    required:
    - inventory_category_name
    type: object
  inventoryCategory.InventoryCategoryResponse:
    properties:
      inventory_categories:
        items:
          $ref: '#/definitions/inventoryCategory.InventoryCategoryDTO'
        type: array
      next_page:
        type: integer
      total_pages:
        type: integer
    type: object
  location.LocationDTO:
    properties:
      location_id:
        type: string
      location_name:
        type: string
      warehouse:
        $ref: '#/definitions/location.WarehouseDto'
    type: object
  location.LocationDTOReq:
    properties:
      location_name:
        type: string
      warehouse_id:
        type: string
    required:
    - location_name
    - warehouse_id
    type: object
  location.WarehouseDto:
    properties:
      warehouse_id:
        type: string
      warehouse_name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is an inventory management server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Inventory Management API
  version: "2.0"
paths:
...
  /inventory-categories:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inventoryCategory.InventoryCategoryDTO'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all inventory categories.
      tags:
      - inventory-category
  /inventory-categories/{page}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: path
        name: page
        required: true
        type: integer
      - description: Limit
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/inventoryCategory.InventoryCategoryResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all inventory categories with pagination.
      tags:
      - inventory-category
  /inventory-category:
    post:
      consumes:
      - application/json
      parameters:
      - description: Inventory Category
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inventoryCategory.InventoryCategoryReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/inventoryCategory.InventoryCategoryDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new inventory category.
      tags:
      - inventory-category
  /inventory-category/{inventoryCategoryId}:
    delete:
      parameters:
      - description: Inventory Category ID
        in: path
        name: inventoryCategoryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new inventory category.
      tags:
      - inventory-category
    get:
      consumes:
      - application/json
      parameters:
      - description: Inventory Category ID
        in: path
        name: inventoryCategoryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventoryCategory.InventoryCategoryDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Gets an inventory category by ID.
      tags:
      - inventory-category
    put:
      consumes:
      - application/json
      parameters:
      - description: Inventory Category
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/inventoryCategory.InventoryCategoryReq'
      - description: Inventory Category ID
        in: path
        name: inventoryCategoryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/inventoryCategory.InventoryCategoryDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Update the inventory category.
      tags:
      - inventory-category
...
  /location:
    post:
      consumes:
      - application/json
      parameters:
      - description: Location
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/location.LocationDTOReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/location.LocationDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new location.
      tags:
      - location
  /location/{locationId}:
    delete:
      parameters:
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete location by id.
      tags:
      - location
    put:
      consumes:
      - application/json
      parameters:
      - description: Location DTO
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/location.LocationDTOReq'
      - description: Location ID
        in: path
        name: locationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/location.LocationDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.Response'
      security:
      - ApiKeyAuth: []
      summary: Update the location.
      tags:
      - location
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"